backend/django_portfolio/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
Pillow
python-dotenv
gunicorn
whitenoise
redis
django-redis
uvicorn
channels
daphne
